requirements.txt
aiohttp==3.6.2
amqp==2.6.1
aniso8601==7.0.0
appdirs==1.4.3
APScheduler==3.6.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
autopep8==1.5.3
Babel==2.6.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.7
billiard==3.6.3.0
bleach==3.2.1
blinker==1.4
blosc==1.9.1
boto==2.49.0
branca==0.4.1
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.7
certifi==2019.11.28
cffi==1.14.3
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coverage==5.3
crypto==1.4.1
cryptography==2.8
cssselect2==0.3.0
dbus-python==1.2.16
decorator==4.0.10
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.0
distro==1.4.0
Django==2.2.12
django-filter==2.4.0
django-paytm==0.1
django-sendsms==0.4
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
dnspython==1.16.0
dredd-hooks==0.2.0
drf-paytm==0.0.3
drfaddons==0.1.4
duplicity==0.8.12.0
elasticsearch==7.9.1
elasticsearch-dsl==7.0.0
email-validator==1.1.1
entrypoints==0.3
enum34==1.1.6
envparse==0.2.0
eventlet==0.28.0
factory-boy==3.0.1
Faker==4.1.3
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.2
Flask-Admin==1.5.6
Flask-BabelEx==0.9.3
Flask-Caching==1.9.0
Flask-CeleryExt==0.3.4
Flask-Compress==1.4.0
Flask-Cors==3.0.9
Flask-Elasticsearch==0.2.5
Flask-GraphQL==2.0.1
Flask-Gravatar==0.4.2
Flask-JWT==0.3.2
Flask-JWT-Extended==3.24.1
flask-jwt-router==0.0.26
Flask-Limiter==1.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
flask-mongoengine==0.9.5
Flask-MySQLdb==0.2.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-PyMongo==2.3.0
flask-redis==0.4.0
Flask-REST-JSONAPI==0.12.6
Flask-Script==2.0.6
Flask-Scrypt==0.1.3.6
Flask-Security-Too==3.4.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
flup==1.0.3
folium==0.11.0
forex-python==1.5
future==0.18.2
geoip2==4.1.0
google-compute-engine==2.8.13
graphene==2.1.8
graphene-sqlalchemy==2.3.0
graphene-sqlalchemy-filter==1.12.1
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==0.4.17
gunicorn==20.0.4
healthcheck==1.3.3
html5lib==1.1
httplib2==0.14.0
humanize==2.6.0
icalendar==4.0.7
identify==1.5.5
idna==2.8
iiswsgi==0.3
importlab==0.5.1
importlib-metadata==1.5.0
iniconfig==1.0.1
ipython-genutils==0.1.0
isort==5.5.4
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==18.0.1
kombu==4.6.11
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
ldap3==2.4.1
limits==1.5.1
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
marrow.mailer==4.0.3
marrow.util==1.2.3
marshmallow==2.15.2
marshmallow-jsonapi==0.23.2
maxminddb==2.0.2
mccabe==0.6.1
mongoengine==0.20.0
monotonic==1.5
more-itertools==4.2.0
MouseInfo==0.1.3
multidict==4.7.6
mypy-extensions==0.4.3
mysql-connector==2.2.9
mysql-connector-python==8.0.21
mysqlclient==2.0.1
Naked==0.1.31
netifaces==0.10.4
networkx==2.5
ninja==1.10.0.post2
nodeenv==1.5.0
nose==1.3.7
numpy==1.19.1
oauth2==1.9.0.post1
oauthlib==3.1.0
objproxies==0.9.4
olefile==0.46
omise==0.9.0
opencv-python==4.4.0.42
packaging==20.3
paramiko==2.6.0
passlib==1.7.2
Paste==3.5.0
PasteDeploy==2.1.1
PasteScript==3.2.0
pathlib2==2.1.0
pathspec==0.8.0
paypalrestsdk==1.13.1
paytm==0.1.8
paytmchecksum==1.7.0
pep8==1.7.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==7.2.0
pipenv==11.9.0
pluggy==0.13.1
pre-commit==2.7.1
promise==2.3
prompt-toolkit==1.0.9
protobuf==3.6.1
psutil==5.5.1
psycopg2==2.8.5
psycopg2-binary==2.8.6
ptyprocess==0.5.1
py==1.9.0
pyasn1==0.4.2
PyAutoGUI==0.9.50
pycairo==1.16.2
pycodestyle==2.6.0
pycountry==20.7.3
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.8
pycups==1.9.73
PyGetWindow==0.0.8
Pygments==2.1.3
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
pymongo==3.11.0
PyMsgBox==1.0.8
PyMySQL==0.10.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
PyPDF2==1.26.0
pyperclip==1.8.0
Pyphen==0.9.5
PyRect==0.1.4
pyRFC3339==1.1
PyScreeze==0.1.26
pytest==6.1.0
pytest-cov==2.10.1
python-apt==2.0.0+ubuntu0.20.4.2
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.14.0
python-geoip==1.2
python-geoip-geolite2==2015.303
python-http-client==3.3.1
python-magic==0.4.18
python-pentabarf-xml==0.20
python3-xlib==0.15
pythonds==1.2.1
PyTweening==1.0.3
pytype==2020.9.29
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
qrcode==6.1
redis==3.5.3
regex==2020.9.27
reportlab==3.5.34
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
Rx==1.6.1
screen-resolution-extra==0.0.0
scrypt==0.8.17
SecretStorage==2.3.1
sendgrid==6.4.7
sentry-sdk==0.18.0
shellescape==3.8.1
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.14.0
speaklater==1.4
SQLAlchemy==1.3.19
SQLAlchemy-Continuum==1.3.11
SQLAlchemy-Utils==0.36.8
sqlparse==0.2.4
sshtunnel==0.1.4
starkbank-ecdsa==1.1.0
stripe==2.54.0
systemd-python==234
text-unidecode==1.3
tinycss2==1.0.2
toml==0.10.1
traitlets==4.3.1
ttystatus==0.38
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vine==1.3.0
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
wcwidth==0.1.7
WeasyPrint==51
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
xhtml2pdf==0.2.4
xkit==0.0.0
yarl==1.6.0
zipp==1.0.0
